room_detection_training/requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pycocotools>=2.0.6
albumentations>=1.3.0

# Scientific computing
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress tracking
tqdm>=4.65.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# YOLO models
ultralytics>=8.1.0

# AWS SageMaker
sagemaker>=2.180.0
boto3>=1.28.0

# Optional: GPU acceleration
# cupy-cuda11x>=12.0.0  # Uncomment for CUDA support
